pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gong-detector"
version = "0.1.0"
description = "A Python project for detecting gong sounds using YAMNet"
authors = [
    {name = "Ethan Frost", email = "ethanfrostbvt@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "tensorflow>=2.13.0",
    "tensorflow-hub>=0.14.0", 
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "yt-dlp>=2023.7.6",
    "ffmpeg-python>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
]

[tool.ruff]
target-version = "py39"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=gong_detector",
    "--cov-report=term-missing",
    "--cov-report=html",
] 